requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
whitenoise==6.8.2
django-environ~=0.11.2
coverage==7.6.9